reduces duplication code
